+++ src/resonaate/physics/constants.py
@@ -25,6 +25,8 @@ RAD2SEC = RAD2DEG * DEG2SEC
 SEC2ARCSEC = 15
 ARCSEC2SEC = 1 / SEC2ARCSEC
 AU2KM = 1.49599 * 10**8  # Astronomical Unit to kilometer
+M2KM = 1.0 / 1000.0
+KM2M = 1000.0
 
 # Physics constants
 SOLAR_FLUX = 1367.0  # Solar flux at Earth's orbit, (W/m^2)